# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIRAMA API",
        "description": "API untuk aplikasi mobile SIRAMA (Sistem Informasi Administrasi Kependudukan). Autentikasi menggunakan Laravel Sanctum Bearer Token.",
        "contact": {
            "email": "admin@sirama.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sirama.tunggulmajid.my.id/api",
            "description": "SIRAMA API Server (Production)"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "SIRAMA API Server (Local)"
        }
    ],
    "paths": {
...
        "/pengajuan": {
            "get": {
                "tags": [
                    "Pengajuan"
                ],
                "summary": "Daftar pengajuan milik warga",
                "description": "Mengembalikan semua pengajuan milik warga yang sedang login, diurutkan terbaru. Mendukung filter berdasarkan status.",
                "operationId": "74ff88069604b97d2b1e5ebc1a6f0c79",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status pengajuan",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "berkas_diterima",
                                "ditolak_desa",
                                "diverifikasi_desa",
                                "ditolak_kecamatan",
                                "diverifikasi_kecamatan",
                                "selesai"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar pengajuan dengan pagination",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PengajuanListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pengajuan/stats": {
            "get": {
                "tags": [
                    "Pengajuan"
                ],
                "summary": "Statistik pengajuan milik warga",
                "description": "Mengembalikan jumlah pengajuan milik warga yang sedang login, dikelompokkan per status beserta labelnya.",
                "operationId": "8a8eb461c0c116d43bfb583c60a2ad93",
                "responses": {
                    "200": {
                        "description": "Statistik pengajuan per status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "statuses": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "status": {
                                                                "type": "string",
                                                                "example": "berkas_diterima"
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "Berkas Diterima"
                                                            },
                                                            "jumlah": {
                                                                "type": "integer",
                                                                "example": 2
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pengajuan/statuses": {
            "get": {
                "tags": [
                    "Pengajuan"
                ],
                "summary": "Daftar semua status pengajuan",
                "description": "Endpoint publik — tidak perlu token. Mengembalikan semua nilai status pengajuan beserta labelnya. Berguna untuk mengisi dropdown filter.",
                "operationId": "c709f98571bcea9e103fda9953c8d6ef",
                "responses": {
                    "200": {
                        "description": "Daftar status pengajuan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "berkas_diterima"
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "Berkas Diterima"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/pengajuan/{id}": {
            "get": {
                "tags": [
                    "Pengajuan"
                ],
                "summary": "Detail satu pengajuan",
                "operationId": "f7a2ccd5c9baa55534800a9ad66a478e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pengajuan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PengajuanResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pengajuan/{id}/status": {
            "get": {
                "tags": [
                    "Pengajuan"
                ],
                "summary": "Cek status pengajuan (polling)",
                "operationId": "51031be5ecf306bc3360500390c95bbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status terkini pengajuan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "diverifikasi_desa"
                                        },
                                        "label_status": {
                                            "type": "string",
                                            "example": "Diverifikasi Desa"
                                        },
                                        "lokasi_dokumen": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nik": {
                        "type": "string",
                        "example": "3277010101900001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "no_whatsapp": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "role": {
                        "type": "string",
                        "example": "warga"
                    },
                    "desa": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "RiwayatStatusResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "diverifikasi_desa"
                    },
                    "catatan": {
                        "type": "string",
                        "example": "Berkas lengkap",
                        "nullable": true
                    },
                    "waktu": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-01T08:00:00+07:00"
                    }
                },
                "type": "object"
            },
            "PengajuanResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "jenis_layanan": {
                        "type": "string",
                        "example": "kia"
                    },
                    "label_layanan": {
                        "type": "string",
                        "example": "KIA (Kartu Identitas Anak)"
                    },
                    "status": {
                        "type": "string",
                        "example": "berkas_diterima",
                        "enum": [
                            "berkas_diterima",
                            "ditolak_desa",
                            "diverifikasi_desa",
                            "ditolak_kecamatan",
                            "diverifikasi_kecamatan",
                            "selesai",
                            "diajukan_kembali"
                        ]
                    },
                    "label_status": {
                        "type": "string",
                        "example": "Berkas Diterima"
                    },
                    "no_whatsapp": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "lokasi_dokumen": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "riwayat_statuses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RiwayatStatusResponse"
                        }
                    }
                },
                "type": "object"
            },
            "PengajuanListResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PengajuanResponse"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "total": {
                                "type": "integer",
                                "example": 25
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum. Contoh: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}